test: tighten legacy-branch assertion to int.TryParse
Addresses a Copilot review comment: the legacy GetInstanceID() branch returns
an int, so int.TryParse reflects the contract precisely (was long.TryParse).
Verified on 2022.3.62f2: test passes.

+++ Tests/Editor/ObjectIdExtensionsTests.cs
@@ -35,7 +35,7 @@ namespace WallstopStudios.DataVisualizer.Tests.Editor
                 );
 #else
                 Assert.IsTrue(
-                    long.TryParse(firstId, out _),
+                    int.TryParse(firstId, out _),
                     "The legacy InstanceID branch must produce a numeric (int) id string."
                 );
 #endif
